Revert "Separate tagvalue controller from tagkey controller"
This reverts commit f67f2ed4177d1f7368c46b136a5e13477370239d.
